requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

gitpython>=3.1.0
openai>=2.16.0

# Tree-sitter for code parsing
tree-sitter>=0.22.0
tree-sitter-language-pack>=0.13.0
